% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{are.encoded.data.and.settings.suitable}
\alias{are.encoded.data.and.settings.suitable}
\alias{are.arg.and.settings.suitable}
\title{check some settings encoded data and settings are suitable for continuing transferring}
\usage{
are.arg.and.settings.suitable(data.encoded)
}
\description{
It checks the sharing for the following criteria:
0. sharing is allowed
1. the encoded data is the same as previously stated in the encoding check
2. encoded data exists
3. encoded data is a data frame
4. the data encoded is character
}
\details{
This is a helper function. It cannot be called directly from any client-side
function.
}
