% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDemoDataDS.R
\name{assignDemoDataDS}
\alias{assignDemoDataDS}
\title{assign some data for testing sharing is working as expected}
\usage{
assignDemoDataDS()
}
\description{
assign some data for testing sharing is working as expected
}
\details{
The function assigns two additional R small object in the global env of a
DataSHIELD server: (1) datashield.mtcars.data and (2) datashield.encrypted.data
}
