% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSharingSettingsDS.R
\name{assignSharingSettingsDS}
\alias{assignSharingSettingsDS}
\title{assign the settings required to share parameters}
\usage{
assignSharingSettingsDS()
}
\description{
assign the settings required to share parameters
}
\details{
The function assigns a variable in the global environment of a DataSHIELD server. A
list of parameters is used in the exchange of parameters. These parameters are to identified names
of values stored on the server temporarily during the exchange. Some other values are used to initialise
some matrices rows and columns.
}
