% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTransfertSettings.R
\name{assignTransferSettingsDS}
\alias{assignTransferSettingsDS}
\title{assigns some settings used to transfer some encoded information}
\usage{
assignTransferSettingsDS(current_row = NULL)
}
\arguments{
\item{current_row}{numeric class representing a row in a dataset, where the transfer stopped}
}
\value{
TRUE
}
\description{
This server function sets some settings specific to transfer of encoded data; i.e.,
a current row.
}
