% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignVariable.R
\name{assignVariable}
\alias{assignVariable}
\title{assigns the name of R-server object to the settings used to transfer some encoded information}
\usage{
assignVariable(var_name = NULL)
}
\arguments{
\item{var_name}{a character string with the name of the server R-server object used int the transfer}
}
\value{
TRUE
}
\description{
This server function sets a settings specific to transfer of encoded data; i.e.,
the name of the R object used.
}
