% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{compute.random.number}
\alias{compute.random.number}
\title{generate a random number}
\usage{
compute.random.number(seed = NULL, min.value = NULL, max.value = NULL)
}
\arguments{
\item{seed}{a integer value to set the seed}

\item{min.value}{the minimum value used in the range}

\item{max.value}{the maximum value used in the range}
}
\value{
a number randomly generated
}
\description{
generate a random number
}
