% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{create.vector}
\alias{create.vector}
\title{convert character string holding semi-columns separated values into a character vector}
\usage{
create.vector(var_names = "")
}
\arguments{
\item{var_names}{- a character vector of one element with ";" separated variable names}
}
\value{
a character vector
}
\description{
This function converts n variables names into a character vector of n character values
}
