% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danger_options.R
\name{danger_options}
\alias{danger_options}
\title{retrieve the settings for the package}
\usage{
danger_options()
}
\description{
This function uses the option "dsSS_settings" and the global enviroment
to retrieve the settings
}
