% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decryptDataDS.R
\name{decryptDataDS}
\alias{decryptDataDS}
\title{decrypt data received from another server}
\usage{
decryptDataDS()
}
\description{
This server function decrypts some received data from a server acting as a "receiver"
}
