% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encoded_data}
\alias{encoded_data}
\title{A dataset containing 32 rows and 12 columns. It has encoded two columns of the mtcars
datasets (i.e.,qsec and hp). The encoding method is as follow:
 \deqn{((values + 15e6)^{0.87}) + 12348223}
 The columns have been obfuscated within a dataframe made of 10 columns. The encoded
 values of hp is column 9 and qsec is column 4.}
\format{
An object of class \code{data.frame} with 32 rows and 10 columns.
}
\source{
\url{https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/mtcars}
}
\usage{
encoded_data
}
\description{
A dataset containing 32 rows and 12 columns. It has encoded two columns of the mtcars
datasets (i.e.,qsec and hp). The encoding method is as follow:
 \deqn{((values + 15e6)^{0.87}) + 12348223}
 The columns have been obfuscated within a dataframe made of 10 columns. The encoded
 values of hp is column 9 and qsec is column 4.
}
\keyword{datasets}
