% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{generate.secure.seeds}
\alias{generate.secure.seeds}
\alias{generate.secure.seed}
\title{uses the footprint set in the Opal server to generate a seed}
\usage{
generate.secure.seed(settings = list())
}
\arguments{
\item{settings}{- the settings list.}
}
\value{
a seed
}
\description{
This function converts n variables names into a character vector of n character values
}
