% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get.name}
\alias{get.name}
\title{read the name from an option}
\usage{
get.name(current.name = "", option.name = "")
}
\arguments{
\item{current.name}{- A character argument representing a default value. Set to ""}

\item{option.name}{- A character argument representing a R option. Set to ""}
}
\description{
This function read the name of an option. If it it does not
exists then the current name value is returned.
}
