% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get.settings.name}
\alias{get.settings.name}
\title{retrieve the name of the settings object}
\usage{
get.settings.name()
}
\description{
This function uses the option "dsSS_settings" to retrieve this information
}
