% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get.settings}
\alias{get.settings}
\title{retrieve the settings for the package}
\usage{
get.settings(envir = globalenv())
}
\arguments{
\item{envir}{the environment set by default to globalenv}
}
\description{
This function uses the option "dsSS_settings" and the global enviroment
to retrieve the settings
}
