% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get.sharing.name}
\alias{get.sharing.name}
\title{retrieve the name of the sharing object}
\usage{
get.sharing.name(envir = globalenv())
}
\arguments{
\item{envir}{the environment set by default to globalenv}
}
\description{
This function uses the option "dsSS_settings" to retrieve this information and the global env
}
