% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get.sharing}
\alias{get.sharing}
\title{retrieve the sharing R object}
\usage{
get.sharing(envir = globalenv())
}
\arguments{
\item{envir}{the environment set by default to globalenv}
}
\value{
the sharing R object if it has been created. Otherwise an empty list.
}
\description{
This function uses the option "dsSS_settings" and the global enviroment
to retrieve the sharing of the data
}
