% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get.transfer.name}
\alias{get.transfer.name}
\title{retrieve the name of the transfer object}
\usage{
get.transfer.name(envir = globalenv())
}
\arguments{
\item{envir}{the environment set by default to globalenv}
}
\value{
name of the transfer object
}
\description{
This function uses the option "dsSS_settings" to retrieve this information and the global env
}
