% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordinatesDS.R
\name{getCoordinatesDS}
\alias{getCoordinatesDS}
\title{Retrieves some data related to some coordinates used to encrypt the parameter}
\usage{
getCoordinatesDS()
}
\value{
a list made of a header, a payload, and four properties[a-d]
}
\description{
This server function retrieves some encrypted coordinates to be passed onto the analysis computer
}
\details{
Some encoded data are transformed into a suitable format to be transferred from a DataSHIELD server to a DataSHIELD client. The property.a indicates
the size of the data at encoding time. Property b is the number of coordinates and property c a timestamp. property d a random number.
}
\seealso{
\link[dsShareServer]{assignDataDS} \link[dsShareServer]{getCoordinatesDS}
}
