% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{A dataset containing 32 rows. It is a basic datasets available to R users}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\source{
\url{https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/mtcars}
}
\usage{
mtcars
}
\description{
A dataset containing 32 rows. It is a basic datasets available to R users
}
\keyword{datasets}
