% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextDS.R
\name{nextDS}
\alias{nextDS}
\title{transfer a certain number of rows from some encoded data.}
\usage{
nextDS(data_encoded = NULL, no.rows = 1000)
}
\arguments{
\item{data_encoded}{character argument representing the name of the encoded data}

\item{no.rows}{positive integer value indicating the number of rows to transfer}
}
\value{
a list made of encoded data
}
\description{
This server-side function transfer the next number of rows
to the client. Only suitably encoded data can be transferred.
}
\seealso{
\link[dsShareServer]{nextDS}, \link[dsShareServer]{isEndOfDataDS}
}
