% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEncryptingDataDS.R
\name{removeEncryptingDataDS}
\alias{removeEncryptingDataDS}
\title{Remove data used to encrypt some parameters}
\usage{
removeEncryptingDataDS(master_mode = TRUE)
}
\description{
This server function can be used to remove the data used to encrypt a parameter between
a two servers
}
