% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTransferDataDS.R
\name{removeTransferDataDS}
\alias{removeTransferDataDS}
\title{Remove data used to exchange some parameters from a DataSHIELD server}
\usage{
removeTransferDataDS()
}
\value{
TRUE if the settings and sharing variables have been deleted. Otherwise, FALSE
}
\description{
This server function deletes the data used in the exchange of the parameters from a DataSHIELD server.
The settings and sharing data structure are deleted. This function is important, to keep temporary
data being analysed by other processes.
}
