% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{are.encoded.data.and.settings.suitable}
\alias{are.encoded.data.and.settings.suitable}
\alias{are.arg.and.settings.suitable}
\title{check some settings encoded data and settings are suitable for continuing transferring}
\usage{
are.arg.and.settings.suitable(data.encoded)
}
\description{
It checks the sharing for the following criteria:
0. sharing is allowed
1. the encoded data is the same as previously stated in the encoding check
2. encoded data exists
3. encoded data is a data frame
4. the data encoded is character
}
\details{
This is a helper function. It cannot be called directly from any client-side
function.
}
\note{
Throws the following errors:
"SERVER::ERR:SHARE::002"  sharing is not allowed or the disclosure setting has not been set to 0 or 1
"SERVER::ERR:SHARE::005"  data.encoded does not exists on the server
"SERVER::ERR:SHARE::008"  data.encoded is not the same R object as previously validated \source{isDataEncodedDS}
"SERVER::ERR:SHARE::009"  data.encoded has yet to be validated by \source{isDataEncodedDS}
"SERVER::ERR:SHARE::010"  data.encoded is not a character vector
}
